frontend/requirements.txt
Flask==3.0.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-firestore==2.14.0
Pillow==10.2.0
